Move registering extensions to a separate file so it can be used in testing
