+++ scripts/core/index.ts
@@ -35,9 +35,8 @@ import ng from 'core/services/ng';
 
 import {reactToAngular1} from 'superdesk-ui-framework';
 
-import {extensions} from 'core/extension-imports.generated';
-import {getSuperdeskApiImplementation} from './get-superdesk-api-implementation';
 import {ExtensionPage} from './extension-page';
+import {registerExtensions} from './register-extensions';
 
 /* globals __SUPERDESK_CONFIG__: true */
 const appConfig = __SUPERDESK_CONFIG__;
@@ -109,39 +108,8 @@ core.run(['$document', ($document) => {
     }
 }]);
 
-core.run(['superdesk', 'modal', 'privileges', (superdesk, modal, privileges) => {
-    Promise.all(
-        Object.keys(extensions).map((extensionId) => {
-            const extensionObject = extensions[extensionId];
-
-            const superdeskApi = getSuperdeskApiImplementation(extensionId, extensions, modal, privileges);
-
-            return extensionObject.extension.activate(superdeskApi).then((activationResult) => {
-                extensionObject.activationResult = activationResult;
-
-                return activationResult;
-            });
-        }),
-    ).then((activationResults) => {
-        const pages = flatMap(activationResults, (activationResult) =>
-            activationResult.contributions != null
-            && activationResult.contributions.pages != null
-                ? activationResult.contributions.pages
-                : [],
-        );
-
-        pages.forEach((page) => {
-            _superdesk
-                .activity(page.url, {
-                    label: page.title,
-                    priority: 100,
-                    category: superdesk.MENU_MAIN,
-                    adminTools: false,
-                    controller: angular.noop,
-                    template: '<sd-extension-page></<sd-extension-page>',
-                });
-        });
-    });
+core.run(['modal', 'privileges', (modal, privileges) => {
+    registerExtensions(_superdesk, modal, privileges);
 }]);
 
 export default core;
